+++ geotrek/trekking/factories.py
@@ -128,3 +128,4 @@ class POIFactory(TopologyFactory):
     name = factory.Sequence(lambda n: u"POI %s" % n)
     description = factory.Sequence(lambda n: u"<p>description %s</p> % n)")
     type = factory.SubFactory(POITypeFactory)
+    published = True